FIX - Issue with Pandas 1.1
Looks like Pandas modified some of their internals lately.  Previously (as recent as 1.0.5),
there would be a new `row` object for every iteration of `pd.DataFrame.apply`.  1.1
re-uses the same object, and if we attempt to modify it in place, then we just
keep getting back the first row over and over.

This commit resolves this issue by explicitly creating a copy the `row`
before applying a function on a dataframe (series `apply` was not affected).